extracted NextStepButton to it's own file
